Make camo environment configurable
We want to set the AWS_REGION for the camo service, but we might want to have more environment varibles in the future as well. If we tell systemd to read from an EnvironmentFile we can modify the deployment process in the future to populate that environment file with AWS_REGION and any other environment variables we need.