Use e2e tests during deployment process
camo is a relativley simple application. With well defined inputs and outputs.

Given a complete set of tests we should be able to attain a very high
certaintiy that the application under test works as expected.

In addition to all the unit-tests we have, with a set of e2e tests we should be able to enable a fully automatted deployment pipeline.